+++ UnrealProj/Source/UnrealProj/Widgets/Popup/Widget_SkillSlot.cpp
@@ -90,7 +90,7 @@ void UWidget_SkillSlot::NativeOnMouseLeave(const FPointerEvent& InMouseEvent)
 
 	auto Char = UGameplayStatics::GetPlayerPawn(GetWorld(), 0);
 	auto MyPlayer = Cast<AMyPlayer>(Char);
-	MyPlayer->GetInstance()->GetUIMgr()->CloseUI(UIType::Information);
+	MyPlayer->GetInstance()->GetUIMgr()->CloseUI((int)UIType::Information);
 
 }
